requirements.txt
alpaca-py>=0.35.0
websockets>=12.0
structlog>=24.0
tenacity>=8.0
numpy>=1.26
streamlit>=1.35.0
plotly>=5.20.0
